skylark-bi-agent/requirements.txt
fastapi
uvicorn
requests
pandas
python-dotenv
openai
